Float evaluators (#218)
* Removed NIU affectedCard parameter from Evaluator.evaluateExpression method

* Refactored Evaluator class to return a float for evaluateExpression method

* Removed ValueSource.evaluateExpression method with affectedCard parameter which was always null

+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/actions/choose/SelectNumberAction.java
@@ -19,8 +19,8 @@ public class SelectNumberAction extends MakeDecisionAction {
 
     public SelectNumberAction(ActionContext context, String choiceText, ValueSource valueSource, String memoryId) {
         super(context.getGame(), context.getPerformingPlayer(), choiceText);
-        _minimum = valueSource.getMinimum(context);
-        _maximum = valueSource.getMaximum(context);
+        _minimum = (int) valueSource.getMinimum(context);
+        _maximum = (int) valueSource.getMaximum(context);
         _actionContext = context;
         _memoryId = memoryId;
     }
